more tidy up
