+++ frontend/app/src/components/home/CurrentChat.svelte
@@ -23,6 +23,7 @@
     import CurrentChatSearchHeader from "./CurrentChatSearchHeader.svelte";
     import GiphySelector from "./GiphySelector.svelte";
     import { messageToForwardStore } from "../../stores/messageToForward";
+    import { toastStore } from "stores/toast";
 
     export let joining: GroupChatSummary | undefined;
     export let chat: ChatSummary;
@@ -141,12 +142,18 @@
         if (!canSend) return;
         let [text, mentioned] = ev.detail;
         if ($currentChatEditingEvent !== undefined) {
-            client.editMessageWithAttachment(
-                chat,
-                text,
-                $currentChatFileToAttach,
-                $currentChatEditingEvent
-            );
+            client
+                .editMessageWithAttachment(
+                    chat,
+                    text,
+                    $currentChatFileToAttach,
+                    $currentChatEditingEvent
+                )
+                .then((success) => {
+                    if (!success) {
+                        toastStore.showFailureToast("errorEditingMessage");
+                    }
+                });
         } else {
             sendMessageWithAttachment(text, mentioned, $currentChatFileToAttach);
         }
